+++ noxfile.py
@@ -36,6 +36,9 @@ _DEFAULT_STORAGE_HOST = "https://storage.googleapis.com"
 
 CURRENT_DIRECTORY = pathlib.Path(__file__).parent.absolute()
 
+# Error if a python version is missing
+nox.options.error_on_missing_interpreters = True
+
 
 @nox.session(python=DEFAULT_PYTHON_VERSION)
 def lint(session):
@@ -75,12 +78,17 @@ def lint_setup_py(session):
     session.run("python", "setup.py", "check", "--restructuredtext", "--strict")
 
 
-def default(session):
+def default(session, install_extras=True):
     constraints_path = str(
         CURRENT_DIRECTORY / "testing" / f"constraints-{session.python}.txt"
     )
     # Install all test dependencies, then install this package in-place.
     session.install("mock", "pytest", "pytest-cov", "brotli", "-c", constraints_path)
+    session.install("mock", "pytest", "pytest-cov", "-c", constraints_path)
+
+    if install_extras:
+        session.install("opentelemetry-api", "opentelemetry-sdk")
+
     session.install("-e", ".", "-c", constraints_path)
 
     # This dependency is included in setup.py for backwards compatibility only
@@ -213,7 +221,7 @@ def cover(session):
     session.run("coverage", "erase")
 
 
-@nox.session(python="3.9")
+@nox.session(python="3.10")
 def docs(session):
     """Build the docs for this library."""